Restructure utility functions
Also some minor work on two_photon.py
